Expanded PhysicsCorrector script
Balls under a threshold velocity are now slowed each update.
The intensity of the slowing is determined by the Slow Factor option.
The velocity at which balls stop moving at all is determined by the Stop
Threshold option.
The velocity at which balls begin to the slowed is determined by the
Slow Threshold option.
Tweaked the stop threshold, slow threshold and slow factor to make the
balls move more naturally.